# api/swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "version": "1.0",
      "title": "Books Web API",
      "description": "Student Assignment API"
    },
    "servers": [
      {
        "url": "http://localhost:3000"
      }
    ],
...
    "paths": {
...
      "/authors": {
        "get": {
          "tags": [
            "Author"
          ],
          "description": "Get list of all authors from database.",
          "summary": "Get authors",
          "operationId": "getAuthors",
          "parameters": [
            {
              "$ref": "#/components/parameters/jsonParam"
            }
          ],
          "responses": {
            "200": {
              "description": "Get all authors successfully.",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Author"
                  }
                }
              }
            },
            "500": {
              "$ref": "#/components/responses/500"
            }
          }
        }
      },
      "/authors/{authorId}": {
        "get": {
          "tags": [
            "Author"
          ],
          "description": "Get info a specific author by ID.",
          "summary": "Get author",
          "operationId": "getAuthor",
          "parameters": [
            {
              "in": "path",
              "name": "authorId",
              "required": true,
              "schema": {
                "type": "string"
              },
              "description": "Auto-generated ID of an author."
            },
            {
              "$ref": "#/components/parameters/jsonParam"
            }
          ],
          "responses": {
            "200": {
              "description": "Get author successfully.",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Author"
                  }
                }
              }
            },
            "404": {
              "$ref": "#/components/responses/404"
            },
            "500": {
              "$ref": "#/components/responses/500"
            }
          }
        },
        "delete": {
          "tags": [
            "Author"
          ],
          "description": "Delete an existing author from database.",
          "summary": "Delete an author",
          "operationId": "deleteAuthor",
          "parameters": [
            {
              "in": "path",
              "name": "authorId",
              "required": true,
              "schema": {
                "type": "string"
              },
              "description": "Auto-generated ID of an author."
            }
          ],
          "responses": {
            "404": {
              "$ref": "#/components/responses/404"
            },
            "500": {
              "$ref": "#/components/responses/500"
            }
          }
        }
      },
      "/authors/new": {
        "post": {
          "tags": [
            "Author"
          ],
          "description": "Add new author into database.",
          "summary": "Add new author",
          "operationId": "postAuthor",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Author"
                }
              }
            }
          },
          "responses": {
            "201": {
              "$ref": "#/components/responses/201"
            },
            "400": {
              "$ref": "#/components/responses/400"
            },
            "500": {
              "$ref": "#/components/responses/500"
            }
          }
        }
      },
      "/authors/{authorId}/edit": {
        "put": {
          "tags": [
            "Author"
          ],
          "description": "Edit data of existing author by Id.",
          "summary": "Edit data",
          "operationId": "putAuthor",
          "parameters": [
            {
              "in": "path",
              "name": "authorId",
              "schema": {
                "type": "string"
              },
              "required": true,
              "description": "Auto-generated ID of an author."
            }
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Author"
                }
              }
            }
          },
          "responses": {
            "400": {
              "$ref": "#/components/responses/400"
            },
            "404": {
              "$ref": "#/components/responses/404"
            },
            "500": {
              "$ref": "#/components/responses/500"
            }
          }
        }
      },
...
      "/register": {
        "post": {
          "tags": [
            "User"
          ],
          "description": "Create a new account.",
          "summary": "Create account",
          "operationId": "postRegister",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/newRegister"
                }
              }
            }
          },
          "responses": {
            "201": {
              "$ref": "#/components/responses/201"
            },
            "405": {
              "$ref": "#/components/responses/405"
            },
            "500": {
              "$ref": "#/components/responses/500"
            }
          }
        }
      },
      "/login": {
        "post": {
          "tags": [
            "User"
          ],
          "description": "Log in an account.",
          "summary": "Login",
          "operationId": "postLogin",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "username": {
                      "type": "string"
                    },
                    "password": {
                      "type": "string",
                      "format": "password"
                    }
                  }
                }
              }
            }
          },
          "responses": {
            "200": {
              "$ref": "#/components/responses/200"
            },
            "405": {
              "$ref": "#/components/responses/405"
            },
            "500": {
              "$ref": "#/components/responses/500"
            }
          }
        }
      },
...
    },
    "components": {
      "schemas": {
...
        "newRegister": {
          "type": "object",
          "properties": {
            "username": {
              "type": "string"
            },
            "password": {
              "type": "string",
              "format": "password"
            },
            "firstName": {
              "type": "string"
            },
            "lastName": {
              "type": "string"
            },
            "gender": {
              "type": "string"
            },
            "address": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string"
                  },
                  "location": {
                    "type": "string"
                  }
                }
              }
            },
            "email": {
              "type": "string",
              "format": "email"
            },
            "country": {
              "type": "string",
              "description": "Auto-generated ID of country."
            },
            "phoneNumber": {
              "type": "string"
            }
          },
          "required": [
            "username",
            "password",
            "firstName",
            "lastName",
            "gender",
            "email",
            "country",
            "phoneNumber"
          ]
        },
        "Author": {
          "type": "object",
          "properties": {
            "firstName": {
              "type": "string"
            },
            "middleName": {
              "type": "string"
            },
            "lastName": {
              "type": "string"
            },
            "origin": {
              "type": "string"
            },
            "gender": {
              "type": "string"
            },
            "dateOfBirth": {
              "type": "string",
              "format": "date-time"
            },
            "biography": {
              "type": "string"
            }
          }
        },
...
      },
      "parameters": {
        "jsonParam": {
          "in": "query",
          "name": "json",
          "schema": {
            "type": "boolean"
          },
          "required": true,
          "description": "If true, return list in json format, otherwise render html page."
        }
      },
      "securitySchemes": {
        "bearerAuth": {
          "type": "http",
          "scheme": "bearer",
          "bearerFormat": "JWT"
        }
      },
      "responses": {
        "200": {
          "description": "Request succeeded."
        },
        "201": {
          "description": "Resource created successfully."
        },
        "400": {
          "description": "Bad input"
        },
        "404": {
          "description": "Specific resource not found."
        },
        "405": {
          "description": "Method not allowed."
        },
        "500": {
          "description": "Internal server error"
        },
...
      }
    }
  }